# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Self-Healing Test Automation Harness API
  description: |
    A comprehensive REST API for managing test execution, monitoring results, and configuring self-healing capabilities.
    
    ## Features
    - **Test Orchestration**: Execute and manage multiple types of tests (unit, e2e, performance, security)
    - **Self-Healing**: AI-powered test recovery and adaptation
    - **Unified Reporting**: Comprehensive test results and analytics
    - **Plugin Architecture**: Extensible test engine support
    - **Real-time Monitoring**: Live test execution status and progress tracking
    
    ## Authentication
    Currently, the API operates without authentication for development purposes. In production, API key authentication will be implemented.
  version: 1.0.0
  contact:
    name: Test Automation Harness Support
    email: support@test-automation-harness.com
    url: https://github.com/your-org/test-automation-harness
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.test-automation-harness.com
    description: Production server
security:
  - ApiKeyAuth: []
paths:
...
  /api/v1/tests/{id}/result:
    get:
      summary: Get complete test result
      description: Get the complete result of a test execution
      tags:
        - Test Execution
      parameters:
        - name: id
          in: path
          required: true
          description: Test ID
          schema:
            type: string
            example: "test_Login_Test_abc123"
        - name: includeArtifacts
          in: query
          description: Include test artifacts
          schema:
            type: boolean
            default: false
        - name: includeHealingAttempts
          in: query
          description: Include healing attempts
          schema:
            type: boolean
            default: true
        - name: includeMetrics
          in: query
          description: Include performance metrics
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: Complete test result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                success: true
                data:
                  testId: "test_Login_Test_abc123"
                  status: "passed"
                  startTime: "2025-01-06T15:30:00.000Z"
                  endTime: "2025-01-06T15:30:30.000Z"
                  duration: 30000
                  engine: "playwright"
                  config: {}
                  results:
                    passed: 5
                    failed: 0
                    skipped: 0
                    total: 5
                  artifacts: {}
                  healingAttempts: []
                  metrics: {}
        '404':
          $ref: '#/components/responses/NotFoundError'
...
  /api/v1/results:
    get:
      summary: Get test results with filtering and pagination
      description: Get test results with various filtering and pagination options
      tags:
        - Test Results
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SortParam'
        - $ref: '#/components/parameters/SortByParam'
        - name: status
          in: query
          description: Filter by status
          schema:
            type: string
            enum: [passed, failed, skipped, running]
        - name: engine
          in: query
          description: Filter by engine name
          schema:
            type: string
        - name: testName
          in: query
          description: Filter by test name
          schema:
            type: string
        - name: startDate
          in: query
          description: Filter by start date
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          description: Filter by end date
          schema:
            type: string
            format: date-time
        - name: includeArtifacts
          in: query
          description: Include artifacts
          schema:
            type: boolean
            default: false
        - name: includeHealingAttempts
          in: query
          description: Include healing attempts
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Test results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                success: true
                data:
                  results:
                    - testId: "test_123"
                      name: "Login Test"
                      status: "passed"
                      engine: "playwright"
                      startTime: "2025-01-06T15:30:00.000Z"
                      endTime: "2025-01-06T15:30:30.000Z"
                      duration: 30000
                      results:
                        passed: 5
                        failed: 0
                        skipped: 0
                        total: 5
                  pagination:
                    page: 1
                    limit: 10
                    total: 25
                    totalPages: 3
                    hasNext: true
                    hasPrev: false
  /api/v1/results/{id}:
    get:
      summary: Get specific test result by ID
      description: Get detailed information about a specific test result
      tags:
        - Test Results
      parameters:
        - name: id
          in: path
          required: true
          description: Test result ID
          schema:
            type: string
            example: "test_123"
      responses:
        '200':
          description: Test result details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                success: true
                data:
                  testId: "test_123"
                  name: "Login Test"
                  status: "passed"
                  engine: "playwright"
                  startTime: "2025-01-06T15:30:00.000Z"
                  endTime: "2025-01-06T15:30:30.000Z"
                  duration: 30000
                  config: {}
                  results:
                    passed: 5
                    failed: 0
                    skipped: 0
                    total: 5
                  artifacts: {}
                  healingAttempts: []
                  metrics: {}
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/v1/results/summary:
    get:
      summary: Get test results summary and statistics
      description: Get summary statistics for test results
      tags:
        - Test Results
      parameters:
        - name: startDate
          in: query
          description: Filter by start date
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          description: Filter by end date
          schema:
            type: string
            format: date-time
        - name: engine
          in: query
          description: Filter by engine name
          schema:
            type: string
        - name: groupBy
          in: query
          description: Group results by field
          schema:
            type: string
            enum: [engine, status, date]
      responses:
        '200':
          description: Test results summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                success: true
                data:
                  total: 100
                  passed: 85
                  failed: 10
                  skipped: 5
                  running: 0
                  successRate: 0.85
                  avgDuration: 25000
                  totalDuration: 2500000
                  byEngine:
                    playwright:
                      total: 50
                      passed: 45
                      failed: 5
                    jest:
                      total: 30
                      passed: 25
                      failed: 5
                    k6:
                      total: 20
                      passed: 15
                      failed: 0
                  byStatus:
                    passed: 85
                    failed: 10
                    skipped: 5
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication (currently optional in development)
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: Items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
    SortParam:
      name: sort
      in: query
      description: Sort order
      schema:
        type: string
        enum: [asc, desc]
        default: desc
    SortByParam:
      name: sortBy
      in: query
      description: Sort field
      schema:
        type: string
        default: createdAt
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          description: Success message
          example: "Operation completed successfully"
        data:
          type: object
          description: Response data
        timestamp:
          type: string
          format: date-time
          description: Response timestamp
          example: "2025-01-06T15:30:00.000Z"
        statusCode:
          type: integer
          description: HTTP status code
          example: 200
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            type:
              type: string
              description: Error type
              example: "ValidationError"
            message:
              type: string
              description: Error message
              example: "Request validation failed"
            statusCode:
              type: integer
              description: HTTP status code
              example: 400
            requestId:
              type: string
              description: Request identifier
              example: "req_abc123"
            details:
              type: string
              description: Additional error details
              example: "Field 'name' is required"
            field:
              type: string
              description: Field that caused the error (for validation errors)
              example: "name"
            value:
              type: string
              description: Invalid value (for validation errors)
              example: ""
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
          example: "2025-01-06T15:30:00.000Z"
  responses:
...
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              type: "NotFoundError"
              message: "Resource not found"
              statusCode: 404
              requestId: "req_abc123"
            timestamp: "2025-01-06T15:30:00.000Z"
...