# src/api/routes/results.ts
/**
 * Test results and reporting API routes for the Self-Healing Test Automation Harness
 * 
 * This module provides REST API endpoints for retrieving test results,
 * generating reports, and accessing test artifacts.
 */

import { Router, Request, Response } from 'express';
import Joi from 'joi';
import { logger } from '../../utils/logger';
import { 
  requestValidationMiddleware, 
  CommonSchemas
} from '../middleware/validation';
import { 
  asyncHandler, 
  createSuccessResponse,
  ApiError,
  NotFoundError 
} from '../middleware/errorHandler';
import { TestResult } from '../../types';

// TODO: Import actual services when they're implemented
// import { ResultsService } from '../../services/ResultsService';
// import { ReportGenerator } from '../../services/ReportGenerator';

const router = Router();

import { testResults, reports } from '../storage/shared';

/**
 * GET /api/v1/results
 * Get test results with filtering and pagination
 */
router.get('/',
  asyncHandler(async (req: Request, res: Response) => {
    const requestId = (req as any).requestId;
    const page = parseInt(req.query['page'] as string || '1', 10);
    const limit = parseInt(req.query['limit'] as string || '10', 10);
    const sort = req.query['sort'] as string || 'desc';
    const sortBy = req.query['sortBy'] as string || 'createdAt';
    const status = req.query['status'] as string;
    const engine = req.query['engine'] as string;
    const testName = req.query['testName'] as string;
    const startDate = req.query['startDate'] as string;
    const endDate = req.query['endDate'] as string;
    const includeArtifacts = req.query['includeArtifacts'] as string;
    const includeHealingAttempts = req.query['includeHealingAttempts'] as string;

    logger.debug('Test results requested', {
      requestId,
      filters: { status, engine, testName, startDate, endDate },
      pagination: { page, limit, sort, sortBy },
    });

    try {
      // TODO: Replace with actual database query
      // const results = await resultsService.getResults({
      //   filters: { status, engine, testName, startDate, endDate },
      //   pagination: { page, limit, sort, sortBy },
      //   includeArtifacts,
      //   includeHealingAttempts,
      // });

      // Demo implementation
      let allResults = Array.from(testResults.values());

      // Apply filters
      if (status) {
        allResults = allResults.filter(r => r.status === status);
      }
      if (engine) {
        allResults = allResults.filter(r => r.name.includes(engine as string));
      }
      if (testName) {
        allResults = allResults.filter(r => r.name.includes(testName as string));
      }
      if (startDate) {
        const start = new Date(startDate as string);
        allResults = allResults.filter(r => r.startTime >= start);
      }
      if (endDate) {
        const end = new Date(endDate as string);
        allResults = allResults.filter(r => r.startTime <= end);
      }

      // Apply sorting
      allResults.sort((a, b) => {
        const aValue = a[sortBy as keyof TestResult];
        const bValue = b[sortBy as keyof TestResult];
        
        if (sort === 'asc') {
          return (aValue as any) > (bValue as any) ? 1 : -1;
        } else {
          return (aValue as any) < (bValue as any) ? 1 : -1;
        }
      });

      // Apply pagination
      const total = allResults.length;
      const startIndex = (page - 1) * limit;
      const endIndex = startIndex + limit;
      const items = allResults.slice(startIndex, endIndex);

      // Remove sensitive data if not requested
      const filteredItems = items.map(result => ({
        ...result,
        ...(includeArtifacts === 'false' && { artifacts: undefined }),
        ...(includeHealingAttempts === 'false' && { healingAttempts: undefined }),
      }));

      res.json(createSuccessResponse({
        items: filteredItems,
        pagination: {
          page,
          limit,
          total,
          totalPages: Math.ceil(total / limit),
          hasNext: endIndex < total,
          hasPrev: page > 1,
        },
        filters: {
          status,
          engine,
          testName,
          startDate,
          endDate,
        },
      }));

    } catch (error) {
      logger.error('Failed to retrieve test results', {
        requestId,
        error: (error as Error).message,
      });

      throw new ApiError(
        'Failed to retrieve test results',
        500,
        true,
        requestId
      );
    }
  })
);

/**
 * GET /api/v1/results/summary
 * Get test results summary and statistics
 */
router.get('/summary',
  asyncHandler(async (req: Request, res: Response) => {
    const requestId = (req as any).requestId;
    const { startDate, endDate, engine, groupBy } = req.query;

    logger.debug('Test results summary requested', {
      requestId,
      startDate,
      endDate,
      engine,
      groupBy,
    });

    try {
      // TODO: Replace with actual database aggregation
      // const summary = await resultsService.getSummary({
      //   startDate,
      //   endDate,
      //   engine,
      //   groupBy,
      // });

      // Demo implementation
      let allResults = Array.from(testResults.values());

      // Apply date filter
      if (startDate) {
        const start = new Date(startDate as string);
        allResults = allResults.filter(r => r.startTime >= start);
      }
      if (endDate) {
        const end = new Date(endDate as string);
        allResults = allResults.filter(r => r.startTime <= end);
      }

      // Apply engine filter
      if (engine) {
        allResults = allResults.filter(r => r.name.includes(engine as string));
      }

      // Calculate summary statistics
      const total = allResults.length;
      const passed = allResults.filter(r => r.status === 'passed').length;
      const failed = allResults.filter(r => r.status === 'failed').length;
      const skipped = allResults.filter(r => r.status === 'skipped').length;
      const running = allResults.filter(r => r.status === 'running').length;

      const successRate = total > 0 ? (passed / total) * 100 : 0;
      const failureRate = total > 0 ? (failed / total) * 100 : 0;

      const avgDuration = total > 0 
        ? allResults.reduce((sum, r) => sum + (r.duration || 0), 0) / total 
        : 0;

      const totalHealingAttempts = allResults.reduce(
        (sum, r) => sum + r.healingAttempts.length, 
        0
      );

      const summary = {
        total,
        passed,
        failed,
        skipped,
        running,
        successRate: Math.round(successRate * 100) / 100,
        failureRate: Math.round(failureRate * 100) / 100,
        avgDuration: Math.round(avgDuration),
        totalHealingAttempts,
        dateRange: {
          startDate: startDate || null,
          endDate: endDate || null,
        },
        groupBy,
        generatedAt: new Date(),
      };

      res.json(createSuccessResponse(summary));

    } catch (error) {
      logger.error('Failed to generate test results summary', {
        requestId,
        error: (error as Error).message,
      });

      throw new ApiError(
        'Failed to generate test results summary',
        500,
        true,
        requestId
      );
    }
  })
);

/**
 * GET /api/v1/results/:id
 * Get specific test result by ID
 */
router.get('/:id',
  asyncHandler(async (req: Request, res: Response) => {
    const requestId = (req as any).requestId;
    const testId = req.params['id'] as string;
    const { 
      includeArtifacts, 
      includeHealingAttempts, 
      includeMetrics, 
      includeOutput 
    } = req.query;

    logger.debug('Specific test result requested', {
      requestId,
      testId,
      includeArtifacts,
      includeHealingAttempts,
      includeMetrics,
      includeOutput,
    });

    const testResult = testResults.get(testId);
    
    if (!testResult) {
      throw new NotFoundError(`Test result with ID ${testId}`, requestId);
    }

    // Filter result based on query parameters
    const result = {
      ...testResult,
      ...(includeArtifacts === 'false' && { artifacts: undefined }),
      ...(includeHealingAttempts === 'false' && { healingAttempts: undefined }),
      ...(includeMetrics === 'false' && { metrics: undefined }),
      ...(includeOutput === 'false' && { output: undefined }),
    };

    res.json(createSuccessResponse(result));

  })
);

/**
 * POST /api/v1/results/reports
 * Generate a test report
 */
router.post('/reports',
  requestValidationMiddleware({
    body: Joi.object({
      name: Joi.string().required().min(1).max(255),
      description: Joi.string().optional().max(1000),
      format: Joi.string().valid('html', 'json', 'xml', 'pdf').default('html'),
      filters: Joi.object({
        testIds: Joi.array().items(Joi.string()).optional(),
        status: Joi.string().valid('passed', 'failed', 'skipped', 'running').optional(),
        engine: Joi.string().optional(),
        startDate: Joi.date().iso().optional(),
        endDate: Joi.date().iso().min(Joi.ref('startDate')).optional(),
      }).optional(),
      options: Joi.object({
        includeArtifacts: Joi.boolean().default(false),
        includeHealingAttempts: Joi.boolean().default(true),
        includeMetrics: Joi.boolean().default(true),
        includeCharts: Joi.boolean().default(true),
        includeDetails: Joi.boolean().default(true),
      }).optional(),
    }),
  }),
  asyncHandler(async (req: Request, res: Response) => {
    const requestId = (req as any).requestId;
    const { name, description, format, filters = {}, options = {} } = req.body;

    logger.info('Test report generation requested', {
      requestId,
      name,
      format,
      filters,
      options,
    });

    try {
      // TODO: Replace with actual report generation
      // const reportGenerator = new ReportGenerator();
      // const report = await reportGenerator.generateReport({
      //   name,
      //   description,
      //   format,
      //   filters,
      //   options,
      // });

      // Demo implementation
      const reportId = generateReportId(name);
      const report = {
        id: reportId,
        name,
        description,
        format,
        status: 'generating',
        createdAt: new Date(),
        filters,
        options,
        downloadUrl: null,
        size: null,
      };

      reports.set(reportId, report);

      // Simulate report generation
      setTimeout(() => {
        const generatedReport = reports.get(reportId);
        if (generatedReport) {
          generatedReport.status = 'completed';
          generatedReport.downloadUrl = `/api/v1/results/reports/${reportId}/download`;
          generatedReport.size = Math.floor(Math.random() * 1000000) + 100000; // 100KB - 1MB
        }
      }, 2000);

      res.status(202).json(createSuccessResponse({
        reportId,
        status: 'generating',
        message: 'Report generation started',
        estimatedCompletion: new Date(Date.now() + 30000), // 30 seconds
      }, 'Report generation started successfully', 202));

    } catch (error) {
      logger.error('Failed to generate test report', {
        requestId,
        error: (error as Error).message,
      });

      throw new ApiError(
        'Failed to generate test report',
        500,
        true,
        requestId
      );
    }
  })
);

/**
 * GET /api/v1/results/reports/:id
 * Get report status and details
 */
router.get('/reports/:id',
  requestValidationMiddleware({
    params: CommonSchemas.idParam,
  }),
  asyncHandler(async (req: Request, res: Response) => {
    const requestId = (req as any).requestId;
    const reportId = req.params['id'] as string;

    logger.debug('Report status requested', {
      requestId,
      reportId,
    });

    const report = reports.get(reportId);
    
    if (!report) {
      throw new NotFoundError(`Report with ID ${reportId}`, requestId);
    }

    res.json(createSuccessResponse(report));

  })
);

/**
 * GET /api/v1/results/reports/:id/download
 * Download generated report
 */
router.get('/reports/:id/download',
  requestValidationMiddleware({
    params: CommonSchemas.idParam,
  }),
  asyncHandler(async (req: Request, res: Response) => {
    const requestId = (req as any).requestId;
    const reportId = req.params['id'] as string;

    logger.debug('Report download requested', {
      requestId,
      reportId,
    });

    const report = reports.get(reportId);
    
    if (!report) {
      throw new NotFoundError(`Report with ID ${reportId}`, requestId);
    }

    if (report.status !== 'completed') {
      throw new ApiError(
        `Report is not ready for download. Status: ${report.status}`,
        409,
        true,
        requestId
      );
    }

    // TODO: Implement actual file download
    // res.download(report.filePath, report.filename);

    // Demo implementation
    res.json(createSuccessResponse({
      reportId,
      downloadUrl: report.downloadUrl,
      filename: `${report.name}.${report.format}`,
      size: report.size,
      contentType: getContentType(report.format),
    }));

  })
);

/**
 * GET /api/v1/results/artifacts/:testId
 * Get test artifacts
 */
router.get('/artifacts/:testId',
  requestValidationMiddleware({
    params: CommonSchemas.idParam,
    query: Joi.object({
      type: Joi.string().valid('screenshot', 'video', 'trace', 'log', 'all').default('all'),
    }),
  }),
  asyncHandler(async (req: Request, res: Response) => {
    const requestId = (req as any).requestId;
    const testId = req.params['testId'] as string;
    const { type } = req.query;

    logger.debug('Test artifacts requested', {
      requestId,
      testId,
      type,
    });

    const testResult = testResults.get(testId);
    
    if (!testResult) {
      throw new NotFoundError(`Test result with ID ${testId}`, requestId);
    }

    let artifacts = testResult.artifacts || [];

    // Filter by type if specified
    if (type !== 'all') {
      artifacts = artifacts.filter(artifact => artifact.type === type);
    }

    res.json(createSuccessResponse({
      testId,
      artifacts,
      total: artifacts.length,
    }));

  })
);

// Helper functions

/**
 * Generate a unique report ID
 */
function generateReportId(name: string): string {
  const timestamp = Date.now().toString(36);
  const random = Math.random().toString(36).substring(2, 8);
  return `report_${name.replace(/[^a-zA-Z0-9]/g, '_')}_${timestamp}_${random}`;
}

/**
 * Get content type for report format
 */
function getContentType(format: string): string {
  const contentTypes = {
    html: 'text/html',
    json: 'application/json',
    xml: 'application/xml',
    pdf: 'application/pdf',
  };
  
  return contentTypes[format as keyof typeof contentTypes] || 'application/octet-stream';
}

export default router;